Some optimizations by hand, add TypeScript declaration
Also update dependencies (for security reasons) and replace uglify-js
with terser.

I intentionally did not modernize any code, so as to preserve browser
compatibility.

---

Optimizations explanation:

Line 1: See line 95.

Line 17: If `json` is not present, it's the fallback anyways. We can
safely remove this value from the checks. Though replacing the `indexOf`
check with multiple checks seems like it would increase size, gzip
easily turns this into a net savings.

Line 95: If window is defined, we'd be using the window variable
anyways. If it wasn't defined (AKA held the value `undefined`), we would
use `undefined`. This can be simplified into just using `window`.
Because we're simply passing the verbatim into the IIFE, we can
eliminate this entirely.